requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-image>=0.21.0
imagecorruptions>=1.1.2
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0
pycocotools>=2.0.6
tqdm>=4.65.0
